1. Модульные тесты. [#159435 #126535]

+++ src/main/java/ru/job4j/converter/Converter.java
@@ -1,11 +1,11 @@
 package ru.job4j.converter;
 
 public class Converter {
-    public static int rubleToEuro(int value) {
+    public static float rubleToEuro(float value) {
         return value / 70;
     }
 
-    public static int rubleToDollar(int value) {
+    public static float rubleToDollar(float value) {
         return value / 60;
     }
 
@@ -18,9 +18,9 @@ public class Converter {
     }
 
     public static void main(String[] args) {
-        int euro = Converter.rubleToEuro(140);
+        float euro = Converter.rubleToEuro(140);
         System.out.println("140 rubles are " + euro + " euro.");
-        int dollar = Converter.rubleToDollar(180);
+        float dollar = Converter.rubleToDollar(180);
         System.out.println("180 rubles are " + dollar + " dollar.");
 
         System.out.println(testResultRubleToDollar(140, 2));
